+++ HISTORY.md
@@ -3,6 +3,8 @@
 ## mlpack ?.?.?
 
 _????-??-??_
+ * Shuffle sequence lengths (#3926)
+
  * Add ability to compile OpenBLAS for windows (#3922)
 
  * Drop pytest-runner and "setup.py test" support (#3921).